+++ ledger/src/blockstore_processor.rs
@@ -1041,7 +1041,7 @@ impl BatchExecutionTiming {
 pub struct ConfirmationProgress {
     pub last_entry: Hash,
     pub tick_hash_count: u64,
-    pub num_shreds: u64,
+    pub num_shreds: u32,
     pub num_entries: usize,
     pub num_txs: usize,
 }
@@ -1104,7 +1104,7 @@ pub fn confirm_slot(
 #[allow(clippy::too_many_arguments)]
 fn confirm_slot_entries(
     bank: &Arc<Bank>,
-    slot_entries_load_result: (Vec<Entry>, u64, bool),
+    slot_entries_load_result: (Vec<Entry>, u32, bool),
     timing: &mut ConfirmationTiming,
     progress: &mut ConfirmationProgress,
     skip_verification: bool,
